requirements.txt
fastapi>=0.109.1
uvicorn[standard]==0.27.0
streamlit==1.31.0
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
openai==1.12.0
pymupdf==1.23.0
faiss-cpu==1.7.4
sentence-transformers==2.3.0
numpy==1.24.0
requests==2.31.0
aiofiles==23.2.0
